Loading post on post page
Post page can load the "post" object from the database when given ?id= in the url.
Also there's a template operation on the database console for manually adding a post.
The loading was probably working before but there was permission problems going on with the database so I changed the permissions